src/main.cpp - ugly windows-way console manipulations
It's unnecessary and ugly, but some Windows console applications wait for any key when exiting...
This breaks the entire paradigm of `CLI` (command-line interface) at the root.
And for no small number of reasons, `system("pause")` is one of the worst evils we can do in this case.
In addition, working with the (user|virtual|etc) terminal is OS-specific.
The task is currently poorly supported by the standard c++ library.
However, to support "backwards compatibility" we can add an interactive mode with `wait_for_any_key_` function (under `ifdef _WIN32` guard) via `::GetStdHandle` & `::ReadConsoleInput` from `WINAPI`.
Note, the quest is not as simple as it might seem at first glance, `wait_for_any_key_` is still not fully integrated with the O.S. specific:
- In essence, this will be reinventing the wheel - not the best task for a pure c++. Better to use `Qt`, `boost` or something for this.
- It is better to determine the owner of the terminal (is our program launched “through a click” or normally).
- Not all terminations will be handled now, it is better to handle real “at exit” case.
- Termination of the process can be performed from another thread with hardcoded 5 sec. timeout!!!
- And there will still be many pitfalls...

P.S: Poor windows application developers... =(